Move the rounding producer from solidmodels/render.jl into Curvilinear so both render paths share it.
GDS rounding becomes a thin wrapper and the orphaned GDS fillet solvers are deleted.
Identical geometry; small fillets over-refine via the marching grid (deferred TODO).
(broke two tests)
